% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{beta_lpdf}
\alias{beta_lpdf}
\title{Log probability density function for the beta distribution}
\usage{
beta_lpdf(x, shape1, shape2)
}
\description{
Log probability density function for the beta distribution
}
\details{
Equivalent to \code{sum(dbeta(x,shape1,shape2,log=TRUE))}. For more information see \code{\link{dbeta}}.
}
