% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{binomial_lpdf}
\alias{binomial_lpdf}
\title{Log probability density function for the beta distribution}
\usage{
binomial_lpdf(x, size, prob)
}
\description{
Log probability density function for the beta distribution
}
\details{
Equivalent to \code{sum(dbinomial(x,size,prob,log=TRUE))}. For more information see \code{\link{dbinom}}.
}
