% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{cauchy_lpdf}
\alias{cauchy_lpdf}
\title{Log probability density function for the cauchy distribution}
\usage{
cauchy_lpdf(x, location, scale)
}
\description{
Log probability density function for the cauchy distribution
}
\details{
Equivalent to \code{sum(dcauchy(x,location,scale,log=TRUE))}. For more information see \code{\link{dcauchy}}.
}
