% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{custom_lpdf}
\alias{custom_lpdf}
\title{Creates custom distributions}
\usage{
custom_lpdf(f)
}
\description{
Creates custom distributions
}
\examples{
\dontrun{
# define normal in terms of precision
f = function(x,mean,prec) dnorm(x, a, 1/sqrt(prec))
new_f = lpdf_custom(f)
}
}
