% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{exponential_lpdf}
\alias{exponential_lpdf}
\title{Log probability density function for the exponential distribution}
\usage{
exponential_lpdf(x, rate)
}
\description{
Log probability density function for the exponential distribution
}
\details{
Equivalent to \code{sum(dexp(x,rate,log=TRUE))}. For more information see \code{\link{dexp}}.
}
