% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{gamma_lpdf}
\alias{gamma_lpdf}
\title{Log probability density function for the gamma distribution}
\usage{
gamma_lpdf(x, shape, scale)
}
\description{
Log probability density function for the gamma distribution
}
\details{
Equivalent to \code{sum(dgamma(x,shape,scale,log=TRUE)))}. For more information see \code{\link{dgamma}}.
}
