% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{geometric_lpdf}
\alias{geometric_lpdf}
\title{Log probability density function for the cauchy distribution}
\usage{
geometric_lpdf(x, prob)
}
\description{
Log probability density function for the cauchy distribution
}
\details{
Equivalent to \code{sum(dgeom(x,prob,log=TRUE))}. For more information see \code{\link{dgeom}}.
}
