% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{log_normal_lpdf}
\alias{log_normal_lpdf}
\title{Log probability density function for the log normal distribution}
\usage{
log_normal_lpdf(x, meanlog, sdlog)
}
\description{
Log probability density function for the log normal distribution
}
\details{
Equivalent to \code{sum(dlnorm(x,meanlog,sdlog,log=TRUE))}. For more information see \code{\link{dlnorm}}.
}
