% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{negative_binomial_lpdf}
\alias{negative_binomial_lpdf}
\title{Log probability density function for the negative binomial distribution}
\usage{
negative_binomial_lpdf(x, size, prob)
}
\description{
Log probability density function for the negative binomial distribution
}
\details{
Equivalent to \code{sum(dnbinom(x,size,prob,log=TRUE))}. For more information see \code{\link{dnbinom}}.
}
