% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{normal_lpdf}
\alias{normal_lpdf}
\title{Log probability density function for the normal distribution}
\usage{
normal_lpdf(x, mean, sd)
}
\description{
Log probability density function for the normal distribution
}
\details{
Equivalent to \code{sum(dnorm(x,mean,sd,log=TRUE))}. For more information see \code{\link{dnorm}}.
}
