% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{poisson_lpdf}
\alias{poisson_lpdf}
\title{Log probability density function for the poisson distribution}
\usage{
poisson_lpdf(x, lambda)
}
\description{
Log probability density function for the poisson distribution
}
\details{
Equivalent to \code{sum(dpois(x,lambda,log=TRUE))}. For more information see \code{\link{dpois}}.
}
