% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{snip}
\alias{snip}
\title{Discard burn-in and/or thin samples}
\usage{
snip(mcmc_list, burnin, thin = 1)
}
\arguments{
\item{mcmc_list}{A \code{mcmc.list} object.}

\item{burnin}{A \code{numeric}. All samples collected below this iteration will be discarded.}

\item{thin}{A \code{numeric}. The thinning interval between consecutive observations.}
}
\value{
A \code{mcmc.list} object
}
\description{
Discard burn-in and/or thin samples
}
