% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{uniform_lpdf}
\alias{uniform_lpdf}
\title{Log probability density function for the uniform distribution}
\usage{
uniform_lpdf(x, min, max)
}
\description{
Log probability density function for the uniform distribution
}
\details{
Equivalent to \code{sum(dunif(x,min,max,log=TRUE))}. For more information see \code{\link{dunif}}.
}
